Added Bootcamp ownership

+++ controllers/bootcamps.js
@@ -61,17 +61,27 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 //  @access Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 
-        const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-            new: true,
-            runValidators: true
-        });
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
-        if (!bootcamp) {
-            return next(
-                new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
-            );
-        }
-        res.status(200).json({ success: true, data: bootcamp });
+    if (!bootcamp) {
+        return next(
+            new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+        );
+    }
+    
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+        );
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    });
+    
+    res.status(200).json({ success: true, data: bootcamp });
 });
 
 // @desc    Delete bootcamp
@@ -86,6 +96,14 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
         );
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`, 401)
+        );
+    }
+
     await Course.deleteMany({ bootcamp: bootcamp._id })
     bootcamp.deleteOne();
     res.status(200).json({ success: true, data: {} });
@@ -120,13 +138,20 @@ exports.getBootcampsInRadius = asyncHandler(async (req, res, next) => {
 
 // @desc    Upload photo for bootcamp
 // @route   PUT /api/v1/bootcamps/:id/photo
-//  @access Private
+// @access  Private
 exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     const bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+        );
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
         );
     }
